requirements.txt
aiosmtpd==1.4.2
alembic==1.7.7
atpublic==3.0.1
attrs==21.4.0
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
langdetect==1.0.9
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
